requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.4
django-bootstrap-icons==0.7.9
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-js-asset==2.0.0
filelock==3.7.0
gunicorn==20.1.0
idna==3.3
mysqlclient==2.1.0
Pillow==9.1.1
platformdirs==2.5.2
psycopg2==2.9.3
python-decouple==3.6
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.14.1
whitenoise==6.2.0
